requirements.txt
flask
flask-cors
pyjwt
fpdf
arabic-reshaper
python-bidi
gunicorn
psycopg2-binary